% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_all.R
\name{cmp_all}
\alias{cmp_all}
\title{Compare counts and labels of a list of dataframes}
\usage{
cmp_all(l, id = "id", include_ids = FALSE, col_groups = c("index",
  "spec"))
}
\arguments{
\item{l}{List of dataframes.}

\item{id}{name of the key variable in the dataframes.}

\item{include_ids}{Logical denoting whether a list column of the ids should be included in the results. The ids in the list show at which values of \code{id} the variable \code{var} contains the value \code{val1}.}

\item{col_groups}{String specifying the order of columns in the resulting dataframe. "index" groups columns by indices. "spec" groups columns by the specification types   c("ex", "varlab", "val", "vallab")}
}
\value{
Dataframe consisting of columns \code{var}, \code{val1}, \code{val2}, ...,  \code{vallab1}, \code{vallab2}, ...,  \code{df1}, \code{df2}, ..., and \code{n}, containing a comparison of the counts of variable values (and their respective value labels) of the dataframes in long format. The indices denote the dataframes in \code{l}. \code{vals_diff} & \code{vallabs_diff} are logical columns indicating if all values, variable / value labels are equal.
}
\description{
Compare counts and labels of a list of dataframes
}
\examples{
# load spss data
path <- system.file("examples", "iris.sav", package = "haven")
df1 <- haven::read_sav(path) \%>\%
  # add id column
  tibble::rownames_to_column("id")

# create a modified copy:
df2 <- df1
df2[1, "Species"] <- 2
# modify the value label of "setosa"
df2$Species <- haven::labelled(df2$Species,
                               labels = c(setosa_mod = 1, versicolor = 2, virginica = 3),
                               label = "Species")

# compare the dataframes counts:
cmp_all(list(df1, df2))
# compare the dataframes and only show the counts where values have changed:
cmp_all(list(df1, df2)) \%>\% dplyr::filter(val1 != val2)
# compare the dataframes and only show the counts where value labels have changed:
cmp_all(list(df1, df2)) \%>\% dplyr::filter(vallab1 != vallab2)

# Create another modified copy
df3 <- df2
df3[2, "Species"] <- 3
# modify the value label of "versicolor"
df2$Species <- haven::labelled(df2$Species,
                               labels = c(setosa = 1, versicolor_mod = 2, virginica = 3),
                               label = "Species_mod")

# compare the dataframes counts:
l <- list(df1, df2, df3)
cmp <- cmp_all(l)
cmp

# compare the dataframes and only show the counts where values have changed:
cmp \%>\% dplyr::filter(val_diff)

# Show where either values or value labels differ:
cmp \%>\% dplyr::filter(val_diff | vallab_diff)
}
