% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_other.R
\name{cmp_cts}
\alias{cmp_cts}
\title{Compare counts of a list of dataframes}
\usage{
cmp_cts(l, id = "id", include_ids = FALSE)
}
\arguments{
\item{l}{List of dataframes.}

\item{id}{name of the key variable in the dataframes.}

\item{include_ids}{Logical denoting whether a list column of the ids should
be included in the results. The ids in the list show at which values of
\code{id} the variable \code{var} contains the numeric value \code{nv1} or character (string) value \code{cv1}.}
}
\value{
Dataframe consisting of columns \code{var}, \code{val}, ...,
\code{vallab1}, \code{vallab2}, ..., & \code{n},
containing a comparison of the counts of variable values (and their
respective value labels) of the two dataframes in long format.
\code{nv_differ} or \code{cv_differ} & \code{vallabs_differ} are logical columns indicating
if all values / value labels are equal.
}
\description{
Compare counts of a list of dataframes
}
\examples{
# load spss data
path <- system.file("examples", "iris.sav", package = "haven")
df1 <- haven::read_sav(path) \%>\%
  # add id column
  tibble::rownames_to_column("id")

# create a modified copy:
df2 <- df1
df2[1, "Species"] <- 2

# compare the dataframes counts:
cmp_cts(list(df1, df2))
# compare the dataframes and only show the counts where values have changed:
cmp_cts(list(df1, df2)) \%>\% dplyr::filter(nv1 != nv2)

# Create another modified copy
df3 <- df2
df3[2, "Species"] <- 3

# compare the dataframes counts:
l <- list(df1, df2, df3)
cmp <- cmp_cts(l)
cmp

# compare the dataframes and only show the counts where values have changed:
cmp \%>\% dplyr::filter(!(nv1 == nv2 & nv2 == nv3))

}
