% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_no_cts_vallabs.R
\name{cmp_no_cts_vallabs}
\alias{cmp_no_cts_vallabs}
\title{Compare counts and labels of a list dataframes}
\usage{
cmp_no_cts_vallabs(l, id = "id")
}
\arguments{
\item{l}{List of dataframes.}

\item{id}{name of the key variable in the dataframes.}
}
\value{
Dataframe consisting of columns \code{var}, \code{val1}, \code{val2}, ...,  \code{vallab1}, \code{vallab2}, ..., containing a comparison of the labelled variable values in the dataframes that don't occur in the data.
}
\description{
Compare counts and labels of a list dataframes
}
\examples{
# load spss data
path <- system.file("examples", "iris.sav", package = "haven")
df1 <- haven::read_sav(path) \%>\%
  # add id column
  tibble::rownames_to_column("id")

# create a modified copy:
df2 <- df1
df2[1, "Species"] <- 2
# modify the value label of "setosa"
df2$Species <- haven::labelled(df2$Species,
                               labels = c(setosa_mod = 1, versicolor = 2, virginica = 3),
                               label = "Species")

# compare the dataframes counts:
cmp_no_cts_vallabs(list(df1, df2))
}
