% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_vars.R
\name{cmp_types}
\alias{cmp_types}
\title{Compare the types of variables in a list of dataframes}
\usage{
cmp_types(l)
}
\arguments{
\item{l}{A list of dataframes.}
}
\value{
A dataframe comparing the types \code{type1},  \code{type2}, ... for
all variables \code{var} in the dataframes in \code{l}. The logical column
\code{type_diff} indicates, if there are differences.
}
\description{
Compare the types of variables in a list of dataframes
}
\examples{
df <- data.frame(fbnr = 1:10,
                 sex = c(2, 1, 2, 1, 1, 2, 2, 1, 2, 1),
                 age = c("24", "23", "23", "41", "23", "39", "30", "18", "31", "48"))
df2 <- df \%>\%
dplyr::mutate(new_var = 1,
              age = as.numeric(age)) \%>\%
  dplyr::select(-sex)
cmp_types(list(df, df2))
}
