% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_all.R
\name{cmp_varlabs}
\alias{cmp_varlabs}
\title{Compare the variable labels between a list of dataframes}
\usage{
cmp_varlabs(l)
}
\arguments{
\item{l}{A list of dataframes.}
}
\value{
A dataframe comparing the \code{varlab}s for all variables in the
dataframes. The logical column \code{varlab_diff} indicates, if there are
differences.
}
\description{
Compare the variable labels between a list of dataframes
}
\examples{
df <- data.frame(fbnr = 1:10,
sex = haven::labelled(c(2, 1, 2, 1, 1, 2, 2, 1, 2, 1),
                      label = "sex",
                      labels = c(MALES = 1, FEMALES = 2)),
                      age = c(24, 23, 23, 41, 23, 39, 30, 18, 31, 48))
df2 <- df \%>\% dplyr::mutate(new_var = 1)
attr(df2$new_var, "label")  <-  "some variable label"
attr(df2$sex, "label")  <-  NULL
cmp_varlabs(list(df, df2))
}
