% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_other.R
\name{cmp_vars}
\alias{cmp_vars}
\title{Compare the existence of variables in a list of dataframes}
\usage{
cmp_vars(l)
}
\arguments{
\item{l}{A list of dataframes.}
}
\value{
A dataframe comparing the existence \code{ex1},  \code{ex2}, ... for
all variables \code{var} in the dataframes in \code{l}. The logical column
\code{ex_diff} indicates, if there are differences.
}
\description{
Compare the existence of variables in a list of dataframes
}
\examples{
df <- data.frame(fbnr = 1:10,
                 sex = c(2, 1, 2, 1, 1, 2, 2, 1, 2, 1),
                 age = c("24", "23", "23", "41", "23", "39", "30", "18", "31", "48"))
df2 <- df \%>\% dplyr::mutate(new_var = 1) \%>\% dplyr::select(-sex)
cmp_vars(list(df, df2))
}
