% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_counts.R
\name{compare_counts}
\alias{compare_counts}
\title{Compare counts and labels of a list dataframes}
\usage{
compare_counts(l, id = "id")
}
\arguments{
\item{l}{List of dataframes.}

\item{id}{name of the key variable in the dataframes.}
}
\value{
Dataframe consisting of 9 columns \code{var}, \code{val}, \code{vallab1}, \code{vallab2}, \code{df1}, \code{df2} & \code{n}, containing a comparison of the counts of variable values (and their respective value labels) of the two dataframes in long format. \code{vals_differ} & \code{vallabs_differ} are logical columns indicating if all values / value labels are equal.
}
\description{
Compare counts and labels of a list dataframes
}
\examples{
# load spss data
path <- system.file("examples", "iris.sav", package = "haven")
df1 <- haven::read_sav(path) \%>\%
  # add id column
  tibble::rownames_to_column("id")

# create a modified copy:
df2 <- df1
df2[1, "Species"] <- 2

# compare the dataframes counts:
compare_counts(list(df1, df2))
# compare the dataframes and only show the counts where values have changed:
compare_counts(list(df1, df2)) \%>\% dplyr::filter(val1 != val2)

# Create another modified copy
df3 <- df2
df3[2, "Species"] <- 3

# compare the dataframes counts:
l <- list(df1, df2, df3)
cmp <- compare_counts(l)
cmp

# compare the dataframes and only show the counts where values have changed:
cmp \%>\% dplyr::filter(vals_differ)

}
