% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longen.R
\name{longen}
\alias{longen}
\title{Create long format of a dataframe, keeping the id column.}
\usage{
longen(df, id = "id")
}
\arguments{
\item{df}{Dataframe with a key variable (\code{id}).}

\item{id}{name of the key variable in the dataframe.}
}
\value{
Dataframe consisting of 3 columns \code{id}, \code{var} & \code{val}, containing the dataframe in long format (based on dplyr::gather).
}
\description{
Create long format of a dataframe, keeping the id column.
}
\examples{
# load spss data
path <- system.file("examples", "iris.sav", package = "haven")
df <- haven::read_sav(path) \%>\%
  # add id column
  tibble::rownames_to_column("id")
df \%>\% longen()
}
