% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_functions.R
\name{make_type_list}
\alias{make_type_list}
\title{Coalesce the nv and cv columns to a new list column conserving the type}
\usage{
make_type_list(df_cts)
}
\arguments{
\item{df_cts}{Data frame}
}
\value{
list column conserving the type
}
\description{
Coalesce the nv and cv columns to a new list column conserving the type
}
