% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_no_cts_vallabs.R
\name{tab_no_cts_vallabs}
\alias{tab_no_cts_vallabs}
\title{Compare value labels not occuring in any of a list of dataframes}
\usage{
tab_no_cts_vallabs(df, id = "id")
}
\arguments{
\item{df}{List of dataframes.}

\item{id}{name of the key variable in the dataframes.}
}
\value{
Dataframe consisting of columns \code{var}, \code{val1}, \code{val2}, ...,  \code{vallab1}, \code{vallab2}, ..., containing a comparison of the labelled variable values in the dataframes that don't occur in the data.
}
\description{
Compare value labels not occuring in any of a list of dataframes
}
\examples{
# load spss data
path <- system.file("examples", "iris.sav", package = "haven")
df1 <- haven::read_sav(path) \%>\%
  # add id column
  tibble::rownames_to_column("id")
# Add some value labels to Sepal.Length
df1$Sepal.Length <- haven::labelled(df1$Sepal.Length,
                                    labels=c(`Sepal.Length = 1` = 1,
                                             `Sepal.Length = 3` = 3,
                                             `Sepal.Length = 5` = 5))

tab_no_cts_vallabs(df1)
}
