% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_vars.R
\name{tab_types}
\alias{tab_types}
\title{Tabulate the variables' types in a dataframe}
\usage{
tab_types(df)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
A dataframe consisting of 2 columns: The \code{var}iable and its
\code{type}.
}
\description{
First \code{unattr()} is being called on df. This should result
in a dataframe of only two types: numeric or character without attributes.
These 2 types are then returned.
}
\examples{
df <- data.frame(fbnr = 1:10,
                 sex = c(2, 1, 2, 1, 1, 2, 2, 1, 2, 1),
                 age = c("24", "23", "23", "41", "23", "39", "30", "18", "31", "48"))
tab_types(df)
}
