% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_varlabs.R
\name{tab_varlabs}
\alias{tab_varlabs}
\title{Create a variable label dataframe}
\usage{
tab_varlabs(df, var = "var", varlab = "varlab")
}
\arguments{
\item{df}{Labelled dataframe, e.g., resulting of the read_sav / read_dta functions  of the package haven.}

\item{var}{Name of the variable column in the resulting dataframe.}

\item{varlab}{Name of the variable label column in the resulting dataframe.}
}
\value{
Dataframe consisting of two columns \code{var} and \code{varlab}, showing all variable labels in \code{df}.
}
\description{
Create a variable label dataframe
}
\examples{
# load spss data
path <- system.file("examples", "iris.sav", package = "haven")
df <- haven::read_sav(path)
attr(df$Sepal.Width, "label") <- "width of the sepal"
tab_varlabs(df)
}
