% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unattr}
\alias{unattr}
\title{Remove attributes from all variables of a dataframe}
\usage{
unattr(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A copy of \code{df} with all attributes removed.
}
\description{
Remove attributes from all variables of a dataframe
}
\examples{
path <- system.file("examples", "iris.sav", package = "haven")
df <- haven::read_sav(path)
df \%>\% unattr()
}
