% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_no_cts_vallabs.R
\name{cmp_no_cts_vallabs}
\alias{cmp_no_cts_vallabs}
\title{Compare value labels not occuring in any of a list of dataframes}
\usage{
cmp_no_cts_vallabs(l, id = "id")
}
\arguments{
\item{l}{List of dataframes.}

\item{id}{name of the key variable in the dataframes.}
}
\value{
Dataframe consisting of columns \code{var}, \code{val1}, \code{val2},
..., \code{ex1}, \code{ex2}, ...,  \code{vallab1}, \code{vallab2}, ...,
containing a comparison of the labelled variable values in the dataframes
that don't occur in the data.
}
\description{
Compare value labels not occuring in any of a list of dataframes
}
\examples{
# load spss data
path <- system.file("examples", "iris.sav", package = "haven")
df1 <- haven::read_sav(path) \%>\%
  # add id column
  tibble::rownames_to_column("id")

# create a modified copy:
df2 <- df1
# Add some value labels to Sepal.Length in both dataframes:
df1$Sepal.Length <- haven::labelled(df1$Sepal.Length,
                                    labels=c(`Sepal.Length = 1` = 1,
                                             `Sepal.Length = 6` = 6,
                                             `Sepal.Length = 9` = 9))
# Add some value labels to Sepal.Length
df2$Sepal.Length <- haven::labelled(df1$Sepal.Length,
                                    labels=c(`Sepal.Length = 1` = 1,
                                             `Sepal.Length = 3` = 3,
                                             `Sepal.Length = 5` = 5))

# compare the dataframes counts:
cmp_no_cts_vallabs(list(df1, df2))
}
