% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_attrs.R
\name{tab_attrs}
\alias{tab_attrs}
\title{Tabulate a dataframe's variables' attributes}
\usage{
tab_attrs(df)
}
\arguments{
\item{df}{Dataframe}
}
\value{
Dataframe with one row per variable in \code{df}, containing the
following columns: "var", "varlab", "val", "vallab", "format.spss",
"labels", "display_width" & "class"
}
\description{
Tabulate a dataframe's variables' attributes
}
\examples{
path <- system.file("examples", "iris.sav", package = "haven")
df <- haven::read_sav(path) \%>\%
  # add id column
  tibble::rownames_to_column("id")

df \%>\% tab_attrs()
# Only show variables with a variable label:
df \%>\% tab_attrs() \%>\% dplyr::filter(!sapply(varlab, is.null))
}
