% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_labs.R
\name{tab_labs}
\alias{tab_labs}
\title{Create a variable & value label dataframe}
\usage{
tab_labs(df, var = "var", varlab = "varlab", val = "val",
  vallab = "vallab")
}
\arguments{
\item{df}{Labelled dataframe, e.g., resulting of the read_sav / read_dta functions  of the package haven.}

\item{var}{Name of the variable column in the resulting dataframe.}

\item{varlab}{Name of the variable label column in the resulting dataframe.}

\item{val}{Name of the value column in the resulting dataframe.}

\item{vallab}{Name of the value label column in the resulting dataframe.}
}
\value{
Dataframe consisting of 4 columns \code{var}, \code{varlab}, \code{val}  and \code{vallab}, showing all variable & value labels in \code{df}.
}
\description{
Create a variable & value label dataframe
}
\examples{
# load spss data
path <- system.file("examples", "iris.sav", package = "haven")
df <- haven::read_sav(path)
tab_labs(df)
}
